requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
google-genai>=1.0.0
python-dotenv>=1.0.1
google-cloud-storage>=2.14.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
rapidfuzz>=3.10.0